Added back some dropped features
- Rough search term support
- Added requirements.txt
- Added feed limit argument
- Renamed `seddit2.py` to `seddit.py`
- Removed `searchCache.py`
- Removed `util.py` functions that are not used by the rewrite
- Added more numbers to `default.csv`
- Moved `example_lists\\` to `term_groups\\`
