+++ filters/default.csv
@@ -1,4 +1,4 @@
 the,of,to,and,a,in,is,it,you,that,he,was,for,on,are,with,as,I,his,they,be,at,one,have,this,from,or,had,by,not,word,but,what,some,we,can,out,other,were,all,there,when,up,use,your,how,said,an,each,she,which,do,their,time,if,will,way,about,many,then,them,write,would,like,so,these,her,long,make,thing,see,him,two,has,look,more,day,could,go,come,did,number,sound,no,most,people,my,over,know,water,than,call,first,who,may,down,side,been,now,find,does,any,just
 me,ve
-1,2,3,4,5,6,7,8,9,0
 b,c,d,e,f,g,h,j,k,l,m,n,o,p,q,s,t,v,w,x,z
+0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99,100
